refactor: mv /exporter/metrics.go to /metrics/metrics_collector.go
